Scala migration

+++ multi-gammon-core/src/main/java/org/akoshterek/backgammon/agent/inputrepresentation/PointCodec.scala
@@ -3,5 +3,5 @@ package org.akoshterek.backgammon.agent.inputrepresentation
 trait PointCodec {
     def getInputsPerPoint: Int = 4
 
-    def setPoint(men: Byte, inputs: Array[Double], offset: Int)
+    def setPoint(men: Int, inputs: Array[Double], offset: Int)
 }
